authenticates with file server

+++ routes/files.js
@@ -11,36 +11,61 @@ const upload = multer({
 const request = require('request');
 
 const repo = require('../repo/Repo.js');
-const settings = require('../settings.json');
+
+let accessToken = null;
+
+authenticateFileServer();
+
+function authenticateFileServer() {
+    request.post({
+        url: process.env.FILE_SERVER_LINK + '/authenticate',
+        headers: {
+            'Content-Type': 'application/json'
+        },
+        body: JSON.stringify({
+            authkey: process.env.FILE_SERVER_AUTHKEY
+        })
+    }, (error, response, body) => {
+        accessToken = JSON.parse(response.body).accessToken;
+    });
+}
 
 router.get('/uploads/:id', (req, res) => {
-    repo.find('uploads', { id: req.params.id })
-        .then(result => {
-            return res.send(result[0]);
-        });
-});
+    if (req.query.filename !== undefined) {
+        repo.find('uploads', { id: req.params.id })
+            .then(result => {
+                const file = result[0].files.find(file => file.originalName === req.query.filename);
+                if (file !== undefined) {
+                    request({
+                        url: process.env.FILE_SERVER_LINK + '/file/' + file.filename,
+                        headers: {
+                            'Authorization': 'Bearer ' + accessToken
+                        }
+                    }, (error, response, body) => {
+                        if (response.statusCode === 403 || response.statusCode === 401) {
+                            authenticateFileServer();
+                            return res.send({ error: 'Error authenticating, try again.' })
+                        } else {
+                            const data = new Uint8Array(JSON.parse(response.body).file.data);
 
-router.get('/uploads/download/:id/:filename', (req, res) => {
-    repo.find('uploads', { id: req.params.id })
-        .then(result => {
-            const file = result[0].files.find(file => file.originalName === req.params.filename);
-            if (file !== undefined) {
-                request({
-                    url: settings.fileServerLink + '/file/' + file.filename
-                }, (error, response, body) => {
-                    const data = new Uint8Array(JSON.parse(response.body).file.data);
-
-                    fs.writeFileSync(path.join(__dirname, '../temp', file.originalName), data);
-                    res.download(path.join(__dirname, '../temp', file.originalName), (error) => {
-                        if (!error) {
-                            fs.unlinkSync(path.join(__dirname, '../temp', file.originalName));
+                            fs.writeFileSync(path.join(__dirname, '../temp', file.originalName), data);
+                            res.download(path.join(__dirname, '../temp', file.originalName), (error) => {
+                                if (!error) {
+                                    fs.unlinkSync(path.join(__dirname, '../temp', file.originalName));
+                                }
+                            });
                         }
                     });
-                });
-            } else {
-                return res.sendStatus(404);
-            }
-        });
+                } else {
+                    return res.sendStatus(404);
+                }
+            });
+    } else {
+        repo.find('uploads', { id: req.params.id })
+            .then(result => {
+                return res.send(result[0]);
+            });
+    }
 });
 
 router.post('/uploads', upload.array('files'), (req, res) => {
@@ -60,33 +85,39 @@ router.post('/uploads', upload.array('files'), (req, res) => {
     });
 
     request.post({
-        url: settings.fileServerLink + '/files',
+        url: process.env.FILE_SERVER_LINK + '/files',
         formData: {
             files: filesReadStreams,
-            filetype: 'file.mimetype',
+            filetype: 'file.mimetype', // maybe don't need filetype & filename
             filename: 'file'
+        },
+        headers: {
+            'Authorization': 'Bearer ' + accessToken
         }
     }, (error, response, body) => {
-        const files = JSON.parse(response.body).files;
+        if (response.statusCode === 403 || response.statusCode === 401) {
+            authenticateFileServer();
+            return res.send({ error: 'Error authenticating, try again.' })
+        } else {
+            const files = JSON.parse(response.body).files;
 
-        req.files.map(file => {
-            fs.unlinkSync(path.join(__dirname, '../temp', file.filename));
-            files.map(file0 => {
-                if (file.filename === file0.originalName) {
-                    fileObject.files.push({
-                        filename: file0.filename,
-                        originalName: file.originalname,
-                        encoding: file.encoding,
-                        mimetype: file.mimetype
-                    });
-                }
+            req.files.map(file => {
+                fs.unlinkSync(path.join(__dirname, '../temp', file.filename));
+                files.map(file0 => {
+                    if (file.filename === file0.originalName) {
+                        fileObject.files.push({
+                            filename: file0.filename,
+                            originalName: file.originalname,
+                        });
+                    }
+                });
             });
-        });
 
-        repo.insert('uploads', fileObject)
-            .then(result => {
-                return res.send({ id });
-            });
+            repo.insert('uploads', fileObject)
+                .then(result => {
+                    return res.send({ id });
+                });
+        }
     });
 });
